requirements.txt
fastapi
uvicorn
beautifulsoup4
requests
python-dotenv
pydantic
redis